+++ src/libstore/include/nix/store/globals.hh
@@ -9,50 +9,11 @@
 #include "nix/store/build/derivation-builder.hh"
 #include "nix/store/local-settings.hh"
 #include "nix/store/store-reference.hh"
+#include "nix/store/worker-settings.hh"
 
 #include "nix/store/config.hh"
 
 namespace nix {
-
-struct MaxBuildJobsSetting : public BaseSetting<unsigned int>
-{
-    MaxBuildJobsSetting(
-        Config * options,
-        unsigned int def,
-        const std::string & name,
-        const std::string & description,
-        const StringSet & aliases = {})
-        : BaseSetting<unsigned int>(def, true, name, description, aliases)
-    {
-        options->addSetting(this);
-    }
-
-    unsigned int parse(const std::string & str) const override;
-};
-
-/**
- * The directory where system configuration files are stored.
- *
- * This is needed very early during initialization, before a main
- * `Settings` object can be constructed.
- */
-const std::filesystem::path & nixConfDir();
-
-/**
- * The path to the system configuration file (`nix.conf`).
- */
-static inline std::filesystem::path nixConfFile()
-{
-    return nixConfDir() / "nix.conf";
-}
-
-/**
- * A list of user configuration files to load.
- *
- * This is needed very early during initialization, before a main
- * `Settings` object can be constructed.
- */
-const std::vector<std::filesystem::path> & nixUserConfFiles();
 
 struct LogFileSettings : public virtual Config
 {
@@ -89,7 +50,7 @@ public:
         {"build-compress-log"}};
 };
 
-class Settings : public virtual Config, private LocalSettings, private LogFileSettings
+class Settings : public virtual Config, private LocalSettings, private LogFileSettings, private WorkerSettings
 {
     StringSet getDefaultSystemFeatures();
 
@@ -129,6 +90,19 @@ public:
         return *this;
     }
 
+    /**
+     * Get the worker settings.
+     */
+    WorkerSettings & getWorkerSettings()
+    {
+        return *this;
+    }
+
+    const WorkerSettings & getWorkerSettings() const
+    {
+        return *this;
+    }
+
     static unsigned int getDefaultCores();
 
     /**
@@ -162,85 +136,10 @@ public:
 
     Setting<bool> keepFailed{this, false, "keep-failed", "Whether to keep temporary directories of failed builds."};
 
-    Setting<bool> keepGoing{
-        this, false, "keep-going", "Whether to keep building derivations when another build fails."};
-
-    Setting<bool> tryFallback{
-        this,
-        false,
-        "fallback",
-        R"(
-          If set to `true`, Nix falls back to building from source if a
-          binary substitute fails. This is equivalent to the `--fallback`
-          flag. The default is `false`.
-        )",
-        {"build-fallback"}};
-
     /**
      * Whether to show build log output in real time.
      */
     bool verboseBuild = true;
-
-    Setting<size_t> logLines{
-        this,
-        25,
-        "log-lines",
-        "The number of lines of the tail of "
-        "the log to show if a build fails."};
-
-    MaxBuildJobsSetting maxBuildJobs{
-        this,
-        1,
-        "max-jobs",
-        R"(
-          Maximum number of jobs that Nix tries to build locally in parallel.
-
-          The special value `auto` causes Nix to use the number of CPUs in your system.
-          Use `0` to disable local builds and directly use the remote machines specified in [`builders`](#conf-builders).
-          This doesn't affect derivations that have [`preferLocalBuild = true`](@docroot@/language/advanced-attributes.md#adv-attr-preferLocalBuild), which are always built locally.
-
-          > **Note**
-          >
-          > The number of CPU cores to use for each build job is independently determined by the [`cores`](#conf-cores) setting.
-
-          <!-- TODO(@fricklerhandwerk): would be good to have those shorthands for common options as part of the specification -->
-          The setting can be overridden using the `--max-jobs` (`-j`) command line switch.
-        )",
-        {"build-max-jobs"}};
-
-    Setting<unsigned int> maxSubstitutionJobs{
-        this,
-        16,
-        "max-substitution-jobs",
-        R"(
-          This option defines the maximum number of substitution jobs that Nix
-          tries to run in parallel. The default is `16`. The minimum value
-          one can choose is `1` and lower values are interpreted as `1`.
-        )",
-        {"substitution-max-jobs"}};
-
-    Setting<unsigned int> buildCores{
-        this,
-        0,
-        "cores",
-        R"(
-          Sets the value of the `NIX_BUILD_CORES` environment variable in the [invocation of the `builder` executable](@docroot@/store/building.md#builder-execution) of a derivation.
-          The `builder` executable can use this variable to control its own maximum amount of parallelism.
-
-          <!--
-          FIXME(@fricklerhandwerk): I don't think this should even be mentioned here.
-          A very generic example using `derivation` and `xargs` may be more appropriate to explain the mechanism.
-          Using `mkDerivation` as an example requires being aware of that there are multiple independent layers that are completely opaque here.
-          -->
-          For instance, in Nixpkgs, if the attribute `enableParallelBuilding` for the `mkDerivation` build helper is set to `true`, it passes the `-j${NIX_BUILD_CORES}` flag to GNU Make.
-
-          If set to `0`, nix will detect the number of CPU cores and pass this number via `NIX_BUILD_CORES`.
-
-          > **Note**
-          >
-          > The number of parallel local Nix build jobs is independently controlled with the [`max-jobs`](#conf-max-jobs) setting.
-        )",
-        {"build-cores"}};
 
     /**
      * Read-only mode.  Don't copy stuff to the store, don't change
@@ -277,219 +176,6 @@ public:
           configuration option is set as the empty string.
         )"};
 
-    Setting<time_t> maxSilentTime{
-        this,
-        0,
-        "max-silent-time",
-        R"(
-          This option defines the maximum number of seconds that a builder can
-          go without producing any data on standard output or standard error.
-          This is useful (for instance in an automated build system) to catch
-          builds that are stuck in an infinite loop, or to catch remote builds
-          that are hanging due to network problems. It can be overridden using
-          the `--max-silent-time` command line switch.
-
-          The value `0` means that there is no timeout. This is also the
-          default.
-        )",
-        {"build-max-silent-time"}};
-
-    Setting<time_t> buildTimeout{
-        this,
-        0,
-        "timeout",
-        R"(
-          This option defines the maximum number of seconds that a builder can
-          run. This is useful (for instance in an automated build system) to
-          catch builds that are stuck in an infinite loop but keep writing to
-          their standard output or standard error. It can be overridden using
-          the `--timeout` command line switch.
-
-          The value `0` means that there is no timeout. This is also the
-          default.
-        )",
-        {"build-timeout"}};
-
-    Setting<Strings> buildHook{
-        this,
-        {"nix", "__build-remote"},
-        "build-hook",
-        R"(
-          The path to the helper program that executes remote builds.
-
-          Nix communicates with the build hook over `stdio` using a custom protocol to request builds that cannot be performed directly by the Nix daemon.
-          The default value is the internal Nix binary that implements remote building.
-
-          > **Important**
-          >
-          > Change this setting only if you really know what you’re doing.
-        )"};
-
-    Setting<std::string> builders{
-        this,
-        "@" + (nixConfDir() / "machines").string(),
-        "builders",
-        R"(
-          A semicolon- or newline-separated list of build machines.
-
-          In addition to the [usual ways of setting configuration options](@docroot@/command-ref/conf-file.md), the value can be read from a file by prefixing its absolute path with `@`.
-
-          > **Example**
-          >
-          > This is the default setting:
-          >
-          > ```
-          > builders = @/etc/nix/machines
-          > ```
-
-          Each machine specification consists of the following elements, separated by spaces.
-          Only the first element is required.
-          To leave a field at its default, set it to `-`.
-
-          1. The URI of the remote store in the format `ssh://[username@]hostname[:port]`.
-
-             > **Example**
-             >
-             > `ssh://nix@mac`
-
-             For backward compatibility, `ssh://` may be omitted.
-             The hostname may be an alias defined in `~/.ssh/config`.
-
-          2. A comma-separated list of [Nix system types](@docroot@/development/building.md#system-type).
-             If omitted, this defaults to the local platform type.
-
-             > **Example**
-             >
-             > `aarch64-darwin`
-
-             It is possible for a machine to support multiple platform types.
-
-             > **Example**
-             >
-             > `i686-linux,x86_64-linux`
-
-          3. The SSH identity file to be used to log in to the remote machine.
-             If omitted, SSH uses its regular identities.
-
-             > **Example**
-             >
-             > `/home/user/.ssh/id_mac`
-
-          4. The maximum number of builds that Nix executes in parallel on the machine.
-             Typically this should be equal to the number of CPU cores.
-
-          5. The “speed factor”, indicating the relative speed of the machine as a positive integer.
-             If there are multiple machines of the right type, Nix prefers the fastest, taking load into account.
-
-          6. A comma-separated list of supported [system features](#conf-system-features).
-
-             A machine is only used to build a derivation if all the features in the derivation's [`requiredSystemFeatures`](@docroot@/language/advanced-attributes.html#adv-attr-requiredSystemFeatures) attribute are supported by that machine.
-
-          7. A comma-separated list of required [system features](#conf-system-features).
-
-             A machine is only used to build a derivation if all of the machine’s required features appear in the derivation’s [`requiredSystemFeatures`](@docroot@/language/advanced-attributes.html#adv-attr-requiredSystemFeatures) attribute.
-
-          8. The (base64-encoded) public host key of the remote machine.
-             If omitted, SSH uses its regular `known_hosts` file.
-
-             The value for this field can be obtained via `base64 -w0`.
-
-          > **Example**
-          >
-          > Multiple builders specified on the command line:
-          >
-          > ```console
-          > --builders 'ssh://mac x86_64-darwin ; ssh://beastie x86_64-freebsd'
-          > ```
-
-          > **Example**
-          >
-          > This specifies several machines that can perform `i686-linux` builds:
-          >
-          > ```
-          > nix@scratchy.labs.cs.uu.nl i686-linux /home/nix/.ssh/id_scratchy 8 1 kvm
-          > nix@itchy.labs.cs.uu.nl    i686-linux /home/nix/.ssh/id_scratchy 8 2
-          > nix@poochie.labs.cs.uu.nl  i686-linux /home/nix/.ssh/id_scratchy 1 2 kvm benchmark
-          > ```
-          >
-          > However, `poochie` only builds derivations that have the attribute
-          >
-          > ```nix
-          > requiredSystemFeatures = [ "benchmark" ];
-          > ```
-          >
-          > or
-          >
-          > ```nix
-          > requiredSystemFeatures = [ "benchmark" "kvm" ];
-          > ```
-          >
-          > `itchy` cannot do builds that require `kvm`, but `scratchy` does support such builds.
-          > For regular builds, `itchy` is preferred over `scratchy` because it has a higher speed factor.
-
-          For Nix to use substituters, the calling user must be in the [`trusted-users`](#conf-trusted-users) list.
-
-          > **Note**
-          >
-          > A build machine must be accessible via SSH and have Nix installed.
-          > `nix` must be available in `$PATH` for the user connecting over SSH.
-
-          > **Warning**
-          >
-          > If you are building via the Nix daemon (default), the Nix daemon user account on the local machine (that is, `root`) requires access to a user account on the remote machine (not necessarily `root`).
-          >
-          > If you can’t or don’t want to configure `root` to be able to access the remote machine, set [`store`](#conf-store) to any [local store](@docroot@/store/types/local-store.html), e.g. by passing `--store /tmp` to the command on the local machine.
-
-          To build only on remote machines and disable local builds, set [`max-jobs`](#conf-max-jobs) to 0.
-
-          If you want the remote machines to use substituters, set [`builders-use-substitutes`](#conf-builders-use-substitutes) to `true`.
-        )",
-        {},
-        false};
-
-    Setting<bool> alwaysAllowSubstitutes{
-        this,
-        false,
-        "always-allow-substitutes",
-        R"(
-          If set to `true`, Nix ignores the [`allowSubstitutes`](@docroot@/language/advanced-attributes.md) attribute in derivations and always attempt to use [available substituters](#conf-substituters).
-        )"};
-
-    Setting<bool> buildersUseSubstitutes{
-        this,
-        false,
-        "builders-use-substitutes",
-        R"(
-          If set to `true`, Nix instructs [remote build machines](#conf-builders) to use their own [`substituters`](#conf-substituters) if available.
-
-          It means that remote build hosts fetch as many dependencies as possible from their own substituters (e.g, from `cache.nixos.org`) instead of waiting for the local machine to upload them all.
-          This can drastically reduce build times if the network connection between the local machine and the remote build host is slow.
-        )"};
-
-    Setting<bool> useSubstitutes{
-        this,
-        true,
-        "substitute",
-        R"(
-          If set to `true` (default), Nix uses binary substitutes if
-          available. This option can be disabled to force building from
-          source.
-        )",
-        {"build-use-substitutes"}};
-
-    Setting<unsigned long> maxLogSize{
-        this,
-        0,
-        "max-build-log-size",
-        R"(
-          This option defines the maximum number of bytes that a builder can
-          write to its stdout/stderr. If the builder exceeds this limit, it’s
-          killed. A value of `0` (the default) means that there is no limit.
-        )",
-        {"build-max-log-size"}};
-
-    Setting<unsigned int> pollInterval{this, 5, "build-poll-interval", "How often (in seconds) to poll for locks."};
-
     Setting<Strings> trustedPublicKeys{
         this,
         {"cache.nixos.org-1:6NCHdD59X431o0gWypbMrAURkbJ16ZPMQFGspcDShjY="},
@@ -518,25 +204,6 @@ public:
           generated using `nix-store --generate-binary-cache-key`. The
           corresponding public key can be distributed to other users, who
           can add it to `trusted-public-keys` in their `nix.conf`.
-        )"};
-
-    Setting<unsigned int> tarballTtl{
-        this,
-        60 * 60,
-        "tarball-ttl",
-        R"(
-          The number of seconds a downloaded tarball is considered fresh. If
-          the cached tarball is stale, Nix checks whether it is still up
-          to date using the ETag header. Nix downloads a new version if
-          the ETag header is unsupported, or the cached ETag doesn't match.
-
-          Setting the TTL to `0` forces Nix to always check if the tarball is
-          up to date.
-
-          Nix caches tarballs in `$XDG_CACHE_HOME/nix/tarballs`.
-
-          Files fetched via `NIX_PATH`, `fetchGit`, `fetchMercurial`,
-          `fetchTarball`, and `fetchurl` respect this TTL.
         )"};
 
     Setting<bool> requireSigs{
@@ -629,27 +296,9 @@ public:
         // Don't document the machine-specific default value
         false};
 
-    Setting<std::vector<StoreReference>> substituters{
-        this,
-        std::vector<StoreReference>{StoreReference::parse("https://cache.nixos.org/")},
-        "substituters",
-        R"(
-          A list of [URLs of Nix stores](@docroot@/store/types/index.md#store-url-format) to be used as substituters, separated by whitespace.
-          A substituter is an additional [store](@docroot@/glossary.md#gloss-store) from which Nix can obtain [store objects](@docroot@/store/store-object.md) instead of building them.
-
-          Substituters are tried based on their priority value, which each substituter can set independently.
-          Lower value means higher priority.
-          The default is `https://cache.nixos.org`, which has a priority of 40.
-
-          At least one of the following conditions must be met for Nix to use a substituter:
-
-          - The substituter is in the [`trusted-substituters`](#conf-trusted-substituters) list
-          - The user calling Nix is in the [`trusted-users`](#conf-trusted-users) list
-
-          In addition, each store path should be trusted as described in [`trusted-public-keys`](#conf-trusted-public-keys)
-        )",
-        {"binary-caches"}};
-
+    // move to daemonsettings in another pass
+    //
+    // we'll add a another parameter to processConnection to thread it through
     Setting<std::set<StoreReference>> trustedSubstituters{
         this,
         {},
@@ -695,55 +344,9 @@ public:
           mismatch if the build isn't reproducible.
         )"};
 
+    // move it out in the 2nd pass
     Setting<bool> printMissing{
         this, true, "print-missing", "Whether to print what paths need to be built or downloaded."};
-
-    Setting<std::string> postBuildHook{
-        this,
-        "",
-        "post-build-hook",
-        R"(
-          Optional. The path to a program to execute after each build.
-
-          This option is only settable in the global `nix.conf`, or on the
-          command line by trusted users.
-
-          When using the nix-daemon, the daemon executes the hook as `root`.
-          If the nix-daemon is not involved, the hook runs as the user
-          executing the nix-build.
-
-            - The hook executes after an evaluation-time build.
-
-            - The hook does not execute on substituted paths.
-
-            - The hook's output always goes to the user's terminal.
-
-            - If the hook fails, the build succeeds but no further builds
-              execute.
-
-            - The hook executes synchronously, and blocks other builds from
-              progressing while it runs.
-
-          The program executes with no arguments. The program's environment
-          contains the following environment variables:
-
-            - `DRV_PATH`
-              The derivation for the built paths.
-
-              Example:
-              `/nix/store/5nihn1a7pa8b25l9zafqaqibznlvvp3f-bash-4.4-p23.drv`
-
-            - `OUT_PATHS`
-              Output paths of the built derivation, separated by a space
-              character.
-
-              Example:
-              `/nix/store/l88brggg9hpy96ijds34dlq4n8fan63g-bash-4.4-p23-dev
-              /nix/store/vch71bhyi5akr5zs40k8h2wqxx69j80l-bash-4.4-p23-doc
-              /nix/store/c5cxjywi66iwn9dcx5yvwjkvl559ay6p-bash-4.4-p23-info
-              /nix/store/scz72lskj03ihkcn42ias5mlp4i4gr1k-bash-4.4-p23-man
-              /nix/store/a724znygmd1cac856j3gfsyvih3lw07j-bash-4.4-p23`.
-        )"};
 
     Setting<bool> useXDGBaseDirectories{
         this,
