+++ src/libstore/store-api.cc
@@ -176,7 +176,7 @@ void Store::addMultipleToStore(PathsSource && pathsToCopy, Activity & act, Repai
                     addToStore(info, *source, repair, checkSigs);
                 } catch (Error & e) {
                     nrFailed++;
-                    if (!settings.keepGoing)
+                    if (!settings.getWorkerSettings().keepGoing)
                         throw e;
                     printMsg(lvlError, "could not copy %s: %s", printStorePath(path), e.what());
                     showProgress();
@@ -407,7 +407,7 @@ StorePathSet Store::queryDerivationOutputs(const StorePath & path)
 
 void Store::querySubstitutablePathInfos(const StorePathCAMap & paths, SubstitutablePathInfos & infos)
 {
-    if (!settings.useSubstitutes)
+    if (!settings.getWorkerSettings().useSubstitutes)
         return;
 
     for (auto & path : paths) {
@@ -463,7 +463,7 @@ void Store::querySubstitutablePathInfos(const StorePathCAMap & paths, Substituta
             }
         }
         if (lastStoresException.has_value()) {
-            if (!settings.tryFallback) {
+            if (!settings.getWorkerSettings().tryFallback) {
                 throw *lastStoresException;
             } else
                 logError(lastStoresException->info());
@@ -473,7 +473,7 @@ void Store::querySubstitutablePathInfos(const StorePathCAMap & paths, Substituta
 
 StorePathSet Store::querySubstitutablePaths(const StorePathSet & paths)
 {
-    if (!settings.useSubstitutes)
+    if (!settings.getWorkerSettings().useSubstitutes)
         return StorePathSet();
 
     StorePathSet remaining;
