+++ src/nix/main.cc
@@ -1,3 +1,5 @@
+#include "nix/cmd/common-eval-args.hh"
+#include "nix/fetchers/fetch-settings.hh"
 #include "nix/util/args/root.hh"
 #include "nix/util/current-process.hh"
 #include "nix/cmd/command.hh"
@@ -384,7 +386,7 @@ void mainWrapped(int argc, char ** argv)
        self-aware. That is, it has to know where it is installed. We
        don't think it's sentient.
      */
-    settings.buildHook.setDefault(
+    settings.getWorkerSettings().buildHook.setDefault(
         Strings{
             getNixBin({}).string(),
             "__build-remote",
@@ -552,10 +554,10 @@ void mainWrapped(int argc, char ** argv)
 
     if (!args.useNet) {
         // FIXME: should check for command line overrides only.
-        if (!settings.useSubstitutes.overridden)
-            settings.useSubstitutes = false;
-        if (!settings.tarballTtl.overridden)
-            settings.tarballTtl = std::numeric_limits<unsigned int>::max();
+        if (!settings.getWorkerSettings().useSubstitutes.overridden)
+            settings.getWorkerSettings().useSubstitutes = false;
+        if (!fetchSettings.tarballTtl.overridden)
+            fetchSettings.tarballTtl = std::numeric_limits<unsigned int>::max();
         if (!fileTransferSettings.tries.overridden)
             fileTransferSettings.tries = 0;
         if (!fileTransferSettings.connectTimeout.overridden)
@@ -563,7 +565,7 @@ void mainWrapped(int argc, char ** argv)
     }
 
     if (args.refresh) {
-        settings.tarballTtl = 0;
+        fetchSettings.tarballTtl = 0;
         settings.ttlNegativeNarInfoCache = 0;
         settings.ttlPositiveNarInfoCache = 0;
     }
