+++ src/nix/nix-build/nix-build.cc
@@ -552,7 +552,10 @@ static void main_nix_build(int argc, char ** argv)
 
         env["NIX_BUILD_TOP"] = env["TMPDIR"] = env["TEMPDIR"] = env["TMP"] = env["TEMP"] = tmpDir.path().string();
         env["NIX_STORE"] = store->storeDir;
-        env["NIX_BUILD_CORES"] = fmt("%d", settings.buildCores ? settings.buildCores : settings.getDefaultCores());
+        env["NIX_BUILD_CORES"] =
+            fmt("%d",
+                settings.getLocalSettings().buildCores ? settings.getLocalSettings().buildCores
+                                                       : settings.getDefaultCores());
 
         DerivationOptions<StorePath> drvOptions;
         try {
